+++ src/main/java/com/davsilvam/pokedecks/services/SerieService.java
@@ -1,5 +1,7 @@
 package com.davsilvam.pokedecks.services;
 
+import com.davsilvam.pokedecks.config.errors.exceptions.ResourceNotFoundException;
+import com.davsilvam.pokedecks.models.Serie;
 import com.davsilvam.pokedecks.models.repositories.SerieRepository;
 import com.davsilvam.pokedecks.services.dtos.SerieResponseDTO;
 import com.davsilvam.pokedecks.services.mappers.SerieMapper;
@@ -7,7 +9,6 @@ import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.UUID;
 
 @Service
 @AllArgsConstructor
@@ -15,7 +16,13 @@ public class SerieService {
     private final SerieRepository serieRepository;
 
     public SerieResponseDTO getSerieById(String id) {
-        return SerieMapper.toDTO(serieRepository.findById(id).orElse(null));
+        Serie serie = serieRepository.findById(id).orElse(null);
+
+        if (serie == null) {
+            throw new ResourceNotFoundException("Serie com ID " + id);
+        }
+
+        return SerieMapper.toDTO(serie);
     }
 
     public List<SerieResponseDTO> getAllSeries() {
@@ -25,6 +32,12 @@ public class SerieService {
     }
 
     public void deleteSerieById(String id) {
+        Serie serie = serieRepository.findById(id).orElse(null);
+
+        if (serie == null) {
+            throw new ResourceNotFoundException("Serie com ID " + id);
+        }
+
         serieRepository.deleteById(id);
     }
 }
